webapp/requirements.txt
fastapi
uvicorn
pydantic
ultralytics
Pillow
requests
